+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Trimming now optional (will skip ffmpeg step)
+* Resizing/cropping only rerenders resizer.webm if the video state has changed
 
 ## [v0.0.1] - 2020-12-30
 * [Associated PR](https://github.com/000png/vapoursynth-gui-for-gifs/pull/1)